seperating api from builder
it doesn't belong together and should have been separated from the beginning already
SIGTERM handling added as a prerequisite for upcoming features
